0020512: EDF 1113 OTHER : Using webkit to display documentation
This is a patch on a fix concerning an unresolved symbol while opening a pdf file from the integrated webkit browser.
The previous fix was useful in the case where the system Qt version was higher than the Salome one.
Now the case is reversed (Salome Qt version is higher than the system one) and the fix must not be used.
Both cases are now handled.